requirements.txt
aiohttp
async
backcall
brotlipy
chardet
contourpy
cycler
dateparser
decorator
exceptiongroup
fonttools
joblib
kiwisolver
lxml
matplotlib
multidict
multitasking
numpy
packaging
pandas
Pillow
pyparsing
pytest
python-binance==1.0.15
python-dateutil==2.8.2
pytz
pyzmq
regex
scikit-learn
scipy
seaborn
threadpoolctl
typing
tzlocal
ujson
websockets
yarl
yfinance
schedule
psycopg2-binary
python-decouple